Created all new disk caching threads.  These wait for the available memory to get below a set threshold and shift buffers over to disk.
fixed a bug in that I should never be accessing anything in the pile of vectors that
keep track of the buffers without holding the bufDirMutex lock.  This includes mucking
with the final bufferInfo structure.  This became _far_ more apparent when I started
adding more things that were accessing it for longer times.